backend/requirements.txt
python-multipart
uvicorn
fastapi
python-jose
passlib
bcrypt==3.2
mysql-connector-python
databases
aiomysql